requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

opencv-python==4.6.0.66
numpy>=1.26.3
Pillow>=10.2.0

torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.1.0

paddlepaddle==2.6.2
paddleocr==2.7.3

pyyaml==6.0.1
openpyxl==3.1.2
pandas==2.2.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4